fix(traverse_dir): collapse bad_file if statement

+++ src/directory.rs
@@ -27,20 +27,19 @@ pub fn traverse_dir(
                 result.file_count += path_result.file_count;
                 result.bytes += path_result.bytes;
             }
-        } else if entry_path.is_file() {
-            if check_if_bad_file_ext(entry_path.extension())
+        } else if entry_path.is_file()
+            && (check_if_bad_file_ext(entry_path.extension())
                 || (!custom_bad_files.is_empty()
-                    && check_if_bad_file(&entry_path, custom_bad_files))
-            {
-                result.file_count += 1;
+                    && check_if_bad_file(&entry_path, custom_bad_files)))
+        {
+            result.file_count += 1;
 
-                if let Ok(metadata) = entry_path.metadata() {
-                    result.bytes += metadata.len()
-                }
+            if let Ok(metadata) = entry_path.metadata() {
+                result.bytes += metadata.len()
+            }
 
-                if delete_files {
-                    delete_file(entry_path);
-                }
+            if delete_files {
+                delete_file(entry_path);
             }
         }
     }
